+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 12972,
+  "packagerPid": 6956,
   "expoServerNgrokUrl": "https://fn-29n.minhchinh97h.tower-client.exp.direct",
   "packagerNgrokUrl": "https://packager.fn-29n.minhchinh97h.tower-client.exp.direct",
-  "ngrokPid": 13756,
+  "ngrokPid": 6340,
   "devToolsPort": 19002,
   "webpackServerPort": null
 }
